Fixing minor bug navbar
